# src/main/java/restapi/springboot_restful_endpoints/entity/Proyek.java
package restapi.springboot_restful_endpoints.entity;

import java.time.LocalDateTime;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import lombok.Data;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter

@Entity
@Data
public class Proyek {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;

    private String namaProyek = "-";
    private String client = "-";
    private LocalDateTime tglMulai;
    private LocalDateTime tglSelesai;
    private String pimpinanProyek = "-";
    private String keterangan = "-";
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;
}